agent/personalizedTripPlanner/pyproject.toml
[project]
name = "personalized-trip-planner"
version = "0.1.0"
description = ""
authors = [
    {name = "Chandan Gowda", email = "chandangowdatk23@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12,<4"
packages = ["personalized_trip_planner"]
dependencies = [
    "google-adk>=1.13.0,<2.0.0",
    "google-genai>=1.32.0,<2.0.0",
    "google-cloud>=0.34.0,<0.35.0",
    "google-cloud-aiplatform>=1.111.0,<2.0.0",
    "google-cloud-storage>=2.19.0",
]

[project.optional-dependencies]
dev = [
    "google-cloud-aiplatform[adk,agent-engines,evaluation]>=1.93.0",
    "pytest>=8.3.5",
    "pytest-asyncio>=0.26.0",
    "google-adk[eval]>=1.5.0",
]
deployment = [
    "absl-py>=2.2.1",
]